Factor out get-consistency-proof command to a file
